requirements.txt
aiogram==3.10.0
pytz==2023.3
python-dotenv==1.0.0